Add support for the `--profile` cargo flag.
The primary change here is simply not passing the `--release`
flag when `--profile` is present, since they conflict.  But
handling `--profile` correctly also requires looking in a
different directory for the produced executables, which this
commit also implements.

This commit also incidentally fixes a bug where only a single
command line argument was being passed on to the cargo build
command.